requirements.txt
aiohttp>=3.7.4
appdirs>=1.4.3
async-timeout>=3.0.1
atomicwrites>=1.3.0
attrs>=19.1.0
black>=19.3b0
cairocffi>=1.0.2
CairoSVG>=2.5.1
cffi>=1.12.3
chardet>=3.0.4
Click>=7.0
cssselect2>=0.2.1
defusedxml>=0.6.0
discord.py>=1.2.3
entrypoints>=0.3
flake8>=3.7.8
idna>=2.8
importlib-metadata>=0.19
loguru>=0.3.2
mccabe>=0.6.1
more-itertools>=7.2.0
multidict>=4.5.2
mypy>=0.720
mypy-extensions>=0.4.1
packaging>=19.1
Pillow>=9.0.1
pluggy>=0.12.0
py>=1.10.0
pycodestyle>=2.5.0
pycparser>=2.19
pyflakes>=2.1.1
pyparsing>=2.4.2
pytest>=5.1.1
python-chess>=0.28.2
six>=1.12.0
SQLAlchemy>=1.3.7
tinycss2>=1.0.2
toml>=0.10.0
typed-ast>=1.4.0
typing-extensions>=3.7.4
wcwidth>=0.1.7
webencodings>=0.5.1
websockets>=9.1
yarl>=1.3.0
zipp>=0.5.2